+++ model_builder/addition/add_panel_http_response_generators.py
@@ -31,7 +31,11 @@ def generate_generic_add_panel_http_response(request, efootprint_class_str: str,
     if request.GET.get("efootprint_id_of_parent_to_link_to"):
         context_data["efootprint_id_of_parent_to_link_to"] = request.GET["efootprint_id_of_parent_to_link_to"]
 
-    return render(request, f"model_builder/side_panels/{template_name}_add.html", context=context_data)
+    http_response = render(request, f"model_builder/side_panels/{template_name}_add.html", context=context_data)
+
+    http_response["HX-Trigger-After-Swap"] = "initDynamicForm"
+
+    return http_response
 
 
 def generate_server_add_panel_http_response(request, model_web: ModelWeb):
@@ -139,7 +143,7 @@ def generate_job_add_panel_http_response(request, model_web: ModelWeb):
     }
     possible_job_types_per_service.update({
         service.efootprint_id: [
-            {"label": FORM_TYPE_OBJECT[job.__name__]['label'], "value": job.__name__} for job in
+            {"label": FORM_TYPE_OBJECT[job.__name__]["label"], "value": job.__name__} for job in
             service.compatible_jobs()]
         for service in model_web.services}
     )
@@ -207,6 +211,8 @@ def generate_usage_pattern_add_panel_http_response(request, model_web: ModelWeb)
     for field in form_sections[1]["fields"]:
         if field["attr_name"] == "devices":
             field["input_type"] = "select"
+            field["options"] = field["unselected"]
+            field["selected" ] = field["unselected"][0]["value"]
 
     usage_pattern_input_values = UsagePatternFromForm.default_values()
     usage_pattern_input_values["initial_usage_journey_volume"] = None